Start docs/README.md as a design document
The seed of the documentation is the design documentation to help us
begin discussing design ideas and start implementing.

This design document will surely be re-organized in the documentation
tree, but the README.md is a good place to put it initially.

The content so far is a brain-dump of the ideas that @astrofrog,
@lneilsen and I talked about in June when we initially came up with the
idea. Note the TODOs.